create test-fieldarith.h
